redesign of a main page, initial work on a preferences windows
