+++ GlobalSettings.xaml
@@ -1,0 +1,21 @@
+﻿<Window x:Class="wslcontrol_gui.GlobalSettings"
+        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+        xmlns:local="clr-namespace:wslcontrol_gui"
+        mc:Ignorable="d"
+        Title="GlobalSettings" Height="450" Width="800">
+    <Grid>
+        <Grid>
+            <Grid.ColumnDefinitions>
+                <ColumnDefinition Width="auto"/>
+                <ColumnDefinition/>
+                <ColumnDefinition/>
+            </Grid.ColumnDefinitions>
+            <TextBlock Grid.Row="3" Margin="5,5,5,5">Default WSL version for new distros:</TextBlock>
+            <RadioButton Grid.Row="3" Grid.Column="1"  Margin="5,6,5,4" Content="WSL1" GroupName="WSLVersionRadio" Name="WSLVersionSelected1" Checked="WSLVersionSelected1_Checked"/>
+            <RadioButton Grid.Row="3" Grid.Column="2"  Margin="5,6,5,4" Content="WSL2" GroupName="WSLVersionRadio" Name="WSLVersionSelected2" Checked="WSLVersionSelected2_Checked"/>
+        </Grid>
+    </Grid>
+</Window>
