+++ SelectedSettings.xaml
@@ -1,0 +1,13 @@
+﻿<Window x:Class="wslcontrol_gui.SelectedSettings"
+        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+        xmlns:local="clr-namespace:wslcontrol_gui"
+        mc:Ignorable="d"
+        Title="SelectedSettings" Height="450" Width="800">
+    <Grid>
+        <Label Visibility="Hidden">Console window will be opened in the background, ignore and don't close it</Label>
+        <Label>Nothing to see here yet</Label>
+    </Grid>
+</Window>
