+++ SelectedSettings.xaml.cs
@@ -1,0 +1,27 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using System.Windows;
+using System.Windows.Controls;
+using System.Windows.Data;
+using System.Windows.Documents;
+using System.Windows.Input;
+using System.Windows.Media;
+using System.Windows.Media.Imaging;
+using System.Windows.Shapes;
+
+namespace wslcontrol_gui
+{
+    /// <summary>
+    /// Interaction logic for SelectedSettings.xaml
+    /// </summary>
+    public partial class SelectedSettings : Window
+    {
+        public SelectedSettings(string selectedDistro, WSLInterface wsli)
+        {
+            InitializeComponent();
+        }
+    }
+}
